requirements.txt
click==7.1.2
dnspython==2.6.1
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
PyJWT==2.9.0
pymongo==4.8.0
python-dotenv==1.0.1
setuptools==72.1.0
Werkzeug==1.0.1
